Adding /api/v4/users/(ids|statuses) to loadtests.
Over last 7 days:
    /api/v4/channels/TOKEN/posts: 77,524
    /api/v4/users/status/ids: 610,054
    /api/v4/users/ids: 17,043

Unfortunately, it looks like entities also call `/api/v4/users/status/ids` periodically when
polling for user statuses (something we already model in the loadtests). At least in the loadtest,
this accounts for the vast majority of the calls to the status endpoint.

So for /api/v4/users/status/ids, I've just added a fractional chance of also calling this endpoint
as part of querying channel, using no real heuristics to arrive at a useful number. (I picked 80%,
tested repeatedly until it fails, based on a observation of switching channels on pre-release.)

For /api/v4/users/ids, I picked a percentage of 22%, again repeated to make this a lower bound.